feat: Remove redundant account tabs
Removes the "CASH ACCOUNT" and "USDT ACCOUNT" tabs from the Deposit and Withdraw screens. These tabs were not fully implemented and are no longer necessary as the application focuses on a single primary deposit/withdrawal method.

This change also removes the `email` field from the user profile update functionality, as it was not being used or displayed.